refactor: cleanup, fix build
