+++ packages/components/password-policies/index.ts
@@ -1,0 +1,2 @@
+export { PasswordPoliciesComponent } from './password-policies.component';
+export { IPasswordPolicy, IPasswordPolicyConfig, EVENT_POLICY_UPDATED } from './password-policy.model'
